requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
Deprecated==1.2.13
django-phonenumber-field==6.3.0
django-timezone-field==5.0
gunicorn==20.1.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
gunicorn
wrapt==1.14.1
asgiref==3.5.2
Django==4.1
django-countries==7.3.2
django-restframework==0.0.1
djangorestframework==3.13.1
Pillow==9.2.0
typing_extensions==4.3.0
tk==0.1.0
stripe
channels==3.0.5
channels-redis==3.4.1
django-livereload==1.7
django-livereload-server==0.4